# Improve the performance of copying data from a slow input to a slow output