**The idea:**
 is to overlap waitings for input and output completion to allow fetching the next portion of data while the previous portion is saved. It would be easy to explain the difference using a timing diagram:
```js
Time:  |-> -> -> -> -> -> -> -> -> -> ->|
/* The original algorithm:*/            |
Input: | Gggg   Ggggg     Ggg   Gggg    |
Output:|     Sss     Sssss   Sss    Ssss|
/* The new algorithm:*/        |        |
Input: | GgggGggggGgg  Gggg    |<PROFIT>|
Output:|     Sss  SssssSss Ssss|        |
```
**Performance testing:**
 by copying 3000 documents (about 39MB JSON data) between two Elasticsearch clusters (both using remote instances). Both algorithms were started at the same time to have the same impact from a network and/or the cluster load.
The existing algorithm:
> 2.94s user 0.53s system 2% cpu 2:12.68 total
> 2.94s user 0.42s system 2% cpu 2:30.21 total

The new algorithm:
> 2.86s user 0.50s system 3% cpu 1:35.33 total
> 2.98s user 0.44s system 3% cpu 1:51.73 total

So, it's about 35% performance boost.

**Drawbacks:**
To deal with complexity, I had to refactor the original algorithm using promises `async/await` NodeJS features. And this led me to the following drawbacks:
1. I don't see an appropriate way to call the `callback` with both `err` and `totalWrites` arguments. So, in an exceptional case, I just pass the `err` argument in the callback.
2. NodeJS barks if the `callback` function has thrown an exception. To be honest, currently, it's not clear for me, what is the default NodeJS' behavior for callback-style algorithms. But, after this fix, NodeJS will threaten to kill the process in case of an unhandled promise rejection.