@ferronrsmith 2018-11-06
Thanks for your contributions. 

All I see is a conversion from using callbacks to using promises. Not sure where exactly the performance boost comes from.