@igelbox 2018-11-06
@ferronrsmith the latest f780c43fb0ebb2b588e0cc5116f33073365f2575 commit brings the performance.
Yeah, I had to refactor the code to use promises, otherwise, I would have to do a tricky stuff with closures.
Anyway, if you have an idea of how to implement this approach using callbacks and closures, please, let me know.