@ferronrsmith 2018-11-07
it's cool,  don't mind your contribution. Just wanted an explanation. I'll run a few tests first then I'll merge thanks.