@ferronrsmith 2018-11-07
Another issue I see with this change. Because you're sending data to the ES concurrently, there will be instances where it suffers from a `Fast producer and slow consumer` issue. Especially id the **refresh_interval** if ES is low.