@igelbox 2018-11-07
Sorry, I didn't get your point about the `Fast producer and slow consumer` issue.
This implementation waits until the previous portion of data has been sent before getting a next one. So, we may have at most 1 portion of data in memory. E.g. in the timing diagram example, the 4th get request `Gggg` will start only after the 2nd send request `Sssss` has been finished:
```js
Input: | GgggGggggGgg  Gggg    |<PROFIT>|
Output:|     Sss  SssssSss Ssss|        |
```
Also, this implementation doesn't try to execute two parallel get requests nor two parallel send requests. It just allows to overlap one "send" request with a "get" one.

Anyway, I was thinking on a previous comment "All I see is a conversion from using callbacks to using promises." and could you please postpone merging this MR a couple days? I believe I have an idea how to implement this algorithm without using Promises, so, I wanna try this.