@ferronrsmith 2018-11-07
Btw, I don’t have a problem with the promise implementation. It works, and from my testing it seems fine