@igelbox 2019-06-10
Sorry for such a delay. A couple of months ago I tried to re-implement this on callbacks rather than promises but with no success. I thought I try once again, but, days are getting warmer (summer is coming) and will to code is getting weaker )
So, if you don't mind yet let's merge this as it is.