@ferronrsmith 2019-06-10
not really seeing any performance gains from this change after retesting. Think we should shelf this for now. Thanks for your contribution