@igelbox 2019-06-15
Weird. Did you try it using two separated ES instances?
I mean if there was just single then it may use all the performance just to a single read or write step so there will be no benefit to overlap team on instance under a high load.

I tried to measure the benefit once again and get a slight improvement if I dump a remote (several thousand kilometers away) ES index into a local Docker-powered ES  ~ 3:44 or so using `master` branch vs 3:27 using my fix (the command was like `./bin/elasticdump --size=3000`):
* The `master`-branch algorithm:
**  6.22s user 1.07s system 3% cpu 3:44.19 total
**  6.00s user 1.17s system 3% cpu 3:49.49 total
* The new algorithm:
**  6.64s user 1.12s system 3% cpu 3:27.61 total
**  6.28s user 1.30s system 3% cpu 3:35.57 total