@ferronrsmith 2019-06-16
Thanks for sticking with this.