@ferronrsmith 2019-07-10
Just to note there were 2 problems with this change.

1. It introduced a race condition, which in turn led to variable leaking (which was resolved by changing scope)
1. All records weren't being processed in all cases since the logic didn't wait until all promises were resolved.
1. Order is no longer guaranteed due to parallel processing.

It's optimization induced by parallel processing, but draw-backs to this is order is no longer guaranteed. Which means this can't be used `historical indexing`, where variations of changes to a single record maybe be a part of a single payload.