@igelbox 2019-07-10
Weird. I don't expect any out of orders. The only thing can overlap is write over read and read over write. Lemme do a static check today. Or, even better, do you have a scenario to reproduce these bugs or make them happen more frequently?