@ferronrsmith 2019-07-10
https://github.com/taskrabbit/elasticsearch-dump/issues/568 (the data isn't index in order)

https://github.com/taskrabbit/elasticsearch-dump/blob/master/test/test.js#L741 (the output `/tmp/out.json`

The order isn't really a problem unless you're doing historical updates. Which isn't really the purpose of this tool. It does have perf benefits, but this is a drawback due to the overlapping