@igelbox 2019-07-11
Well, I figured it out. The issue in the `file` transport actually. Due to `stream` never paused (we pause `metaStream` only) we may get several subsequent `thisGetCallback` calls.
Previous 6.0.2 version were triggering several concurrent `output.set` requests which might cause some race conditions I believe.
New version uses Promise-s (e.g. `promisify(this.input.get)`) so we simply lose all `thisGetCallback` calls except the last one (which brings us an empty array).